check_tree.pl: Protect logind.conf.in against unwanted changes
It is very annoying that every round of check_tree.pl comes up with a patch like:

```
--- a/src/login/logind.conf.in	2019-01-24 19:40:08.476352606 +0100
+++ b/src/login/logind.conf.in	2019-01-24 19:38:25.494351197 +0100
@@ -12,6 +12,8 @@
 # See logind.conf(5) for details.

 [Login]
+#NAutoVTs=6
+#ReserveVT=6
 #KillUserProcesses=@KILL_USER_PROCESSES@
 #KillOnlyUsers=
 #KillExcludeUsers=root
@@ -34,11 +36,3 @@
 #InhibitorsMax=8192
 #SessionsMax=8192
 #UserTasksMax=33%
-
-[Sleep]
-#SuspendState=mem standby freeze
-#SuspendMode=
-#HibernateState=disk
-#HibernateMode=platform shutdown
-#HybridSleepState=disk
-#HybridSleepMode=suspend platform shutdown
```
This file is meant to be edited by sysadmins/users, so putting in
shell masks is no option here.

This patch adds a stage, just for this file, which removes the
additions of entries we know we do not need, and reverts the removal
of elogind specific configuration lines.

Bug: #2
Closes: #2
Signed-off-by: Sven Eden <sven.eden@prydeworx.com>