check_tree.pl : Change order of checking comments and name reverts
Since we protect comments against certain changes, patches like this
pop up:

```
--- a/src/login/logind-user.c	2019-02-20 07:51:47.953093224 +0100
+++ b/src/login/logind-user.c	2019-02-20 07:52:34.805092111 +0100
@@ -402,6 +402,8 @@
 #endif // 1
         /* Save the user data so far, because pam_elogind will read the XDG_RUNTIME_DIR out of it while starting up
          * elogind --user.  We need to do user_save_internal() because we have not "officially" started yet. */
+        /* Save the user data so far, because pam_elogind will read the XDG_RUNTIME_DIR out of it while starting up
+         * elogind --user.  We need to do user_save_internal() because we have not "officially" started yet. */
         user_save_internal(u);

 #if 0 /// elogind does not spawn user instances of systemd
```
The protection here is pam_elogind. Unfortunately the
`elogind --user` reference is false, that line must be fixed.

However, while the deletion of the original line was correctly
reversed, the addition of the substitution was not stopped. Clearly a
bug in check_tree.pl.

This was caused by doing the check for comment changes first, and
name reverts later. Therefore the order has been changes.

Bug: #5
Closes: #5
Signed-off-by: Sven Eden <sven.eden@prydeworx.com>